hydroreport: add OpenSampleFile
Also make MultiSampleReader merge files in sample order.
This means we'll be able to have as many independent sample
files as we like and merge them all together to produce a
final set of samples.

We take care that when sample files aren't overlapping, they
aren't all opened at the same time.