hydroctl: cope better with daylight savings time changes
Instead of recording the start and end time of a slot as a duration from
midnight, use an absolute time of day, and calculate the start
and end time of the slot using that with respect to the current
time.

We define a new `TimeOfDay` type to make this clear.
We also add a new slot kind, "Continuous" to indicate that the relay
should be enabled for the entire duration of the slot, regardless
of how long that is (it may vary during a DST transition).

We also move some tests to use quicktest instead of gocheck (more still to come).

Currently the time zone is hard-coded, but this will change in the future
(although it might be awkward to thread the time zone change dynamically through the code).